Reorganised the project into cogs
