+++ cogs/errors.py
