+++ cogs/music/music_utils/music_exceptions.py
@@ -1,0 +1,13 @@
+from nextcord.ext import commands
+
+
+class VCError(commands.CommandError):
+    """
+    Exception class for connection errors.
+    """
+
+
+class InvalidVC(VCError):
+    """
+    Exception class for invalid VCs.
+    """
